Baseline event2 kernel driver.
git-svn-id: file:///afs/slac/g/pcds/vol2/svn/pcds/trunk/pcds/epics/modules/ev2_driver/current@25935 ececa60f-9f63-0410-97c7-ea05888eb260

+++ driver/evrmemmap.h
@@ -102,7 +102,7 @@ struct EvrQueues {
   struct DBufInfo  dbq[MAX_EVR_DBQ];       /*  8K + 16 */
   long long ewp;                           /*  20 */
   long long dwp;
-  int  fifofull;
+  long long fifofull;
 };
 
 struct MrfErRegs {
